+++ utils/configs/src/f/fireworks2.ts
@@ -121,6 +121,7 @@ const options: ISourceOptions = {
               sync: false,
               startValue: "max",
               destroy: "min",
+              count: 1,
             },
           },
           effect: {