+++ utils/configs/src/f/index.ts
@@ -1,6 +1,7 @@
 import fallingConfetti from "./fallingConfetti.js";
 import fireworks from "./fireworks.js";
 import fireworks2 from "./fireworks2.js";
+import fireworks3 from "./fireworks3.js";
 import fireworksMask from "./fireworksMask.js";
 import fontawesome from "./fontawesome.js";
 import forward from "./forward.js";
@@ -9,6 +10,7 @@ export default {
   fallingConfetti,
   fireworks,
   fireworks2,
+  fireworks3,
   fireworksMask,
   fontawesome,
   forward,