+++ demo/vanilla/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-beta.7](https://github.com/tsparticles/tsparticles/compare/v4.0.0-beta.6...v4.0.0-beta.7) (2026-04-07)
+
+### Bug Fixes
+
+- small minor fixes ([e52179c](https://github.com/tsparticles/tsparticles/commit/e52179c71bab88d54f12006ee2a0c1dcffeb46ba))
+
 # [4.0.0-beta.6](https://github.com/tsparticles/tsparticles/compare/v4.0.0-beta.5...v4.0.0-beta.6) (2026-04-02)
 
 ### Features